Move reduction kernels to reduction module
