+++ sklearn_numba_dpex/common/__init__.py
@@ -1,0 +1,16 @@
+# TODO: many auxilliary kernels in this module might be better optimized and we could
+# benchmark alternative implementations for each of them, that could include
+#    - using 2D or 3D grid of work groups and work items where applicable (e.g. in
+# some of the kernels that take 2D or 3D data as input) rather than using 1D grid. When
+# doing so, one should be especially careful about how the segments of adjacent work
+# items of size preferred_work_group_size_multiple are dispatched especially regarding
+# RW  operations in memory. A wrong dispatch strategy could slash memory bandwith and
+# reduce performance. Using 2D or 3D grid correctly might on the other hand improve
+# performance since it saves costly indexing operations (like //)
+#    - using numba + dpnp to directly leverage kernels that are shipped in dpnp to
+# replace numpy methods.
+# However, in light of our main goal that is bringing a GPU KMeans to scikit-learn, the
+# importance of those TODOs is currently seen as secondary, since the execution time of
+# those kernels is only a small fraction of the total execution time and the
+# improvements that further optimizations can add will only be marginal. There is no
+# doubt, though, that a lot could be learnt about kernel programming in the process.
